# api/openapi.yaml
openapi: 3.1.0
info:
    title: PP Prime Construction API
    version: 0.1.0
    description: |
        สัญญา API ระหว่างหน้าบ้าน (Angular) และหลังบ้านของระบบบริหารงานก่อสร้าง PP Prime Construction

        หลักการ
        - เงินเป็นบาท (number) ทศนิยมได้ 2 ตำแหน่ง
        - วันที่ตามแผนใช้ `date` (YYYY-MM-DD) เวลาที่เกิดเหตุการณ์ใช้ `date-time` (ISO 8601 พร้อมเขตเวลา)
        - สถานะ/ประเภทเป็นรหัสภาษาอังกฤษ หน้าบ้านแปลเป็นข้อความไทยเอง
        - ข้อผิดพลาดตอบเป็น `application/problem+json` (RFC 9457)
        - การคำนวณทางธุรกิจ (ระดับผู้อนุมัติ กำไร-ขาดทุน ไทม์ไลน์ สถานะงวด) ทำที่หลังบ้าน
        - ทุกการเปลี่ยนแปลงข้อมูล หลังบ้านต้องบันทึก Audit Log เอง (หน้าบ้านไม่มี API เขียน Audit Log)
        - สิทธิ์การเข้าถึงตรวจที่หลังบ้านตามบทบาทใน `GET /roles`; ข้อมูลที่บทบาทนั้นไม่มีสิทธิ์เห็นต้องไม่ถูกส่งมา
servers:
    - url: /api
      description: ค่าเริ่มต้น (หน้าบ้านตั้ง apiBaseUrl ได้)
security:
    - bearerAuth: []
...
paths:
...
    /projects:
        get:
            tags: [projects]
            summary: รายการโครงการที่ผู้ใช้มีสิทธิ์เห็น
            operationId: listProjects
            parameters:
                - { name: status, in: query, schema: { $ref: '#/components/schemas/ProjectStatus' } }
                - { name: group, in: query, description: กลุ่มโครงการ (ดู ProjectGroup), schema: { $ref: '#/components/schemas/ProjectGroup' } }
                - { name: q, in: query, description: ค้นหารหัส ชื่อลูกค้า ผู้รับผิดชอบ, schema: { type: string } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/Project' }
        post:
            tags: [projects]
            summary: เปิดโครงการใหม่ (ต้องมีสิทธิ์ project.create)
            description: |
                หลังบ้านออกรหัสโครงการ (รหัสภูมิภาค + ปี พ.ศ. 2 หลัก + ลำดับ 4 หลัก เช่น CR690003) และบันทึก Audit Log
                โครงการใหม่มีสถานะ pending-contract (ยังไม่มีมูลค่า วันที่ และแผนงาน จนกว่าจะบันทึกสัญญาและตั้งค่างานก่อสร้าง)
            operationId: createProject
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ProjectInput' }
            responses:
                '201':
                    description: เปิดโครงการแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Project' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '422': { $ref: '#/components/responses/ValidationError' }
...
    /projects/{code}:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            operationId: getProject
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Project' }
                '404': { $ref: '#/components/responses/NotFound' }
    /projects/{code}/contract:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        post:
            tags: [projects]
            summary: บันทึกสัญญาและเริ่มแผนงาน (ต้องมีสิทธิ์ project.manage)
            description: |
                ใช้กับโครงการสถานะ pending-contract เท่านั้น
                หลังบ้านบันทึกมูลค่า/วันที่ สร้างแผนงานจากแม่แบบให้พอดีกับวันเริ่ม-วันส่งมอบ สร้างงวดงาน เปลี่ยนสถานะเป็น planning และบันทึก Audit Log
                ไทม์ไลน์ยังไม่ถูกสร้างจนกว่าจะตั้งค่างานก่อสร้าง (PUT /projects/{code}/setup)
                endpoint ข้อมูลรายโครงการ (timeline, installments, photos, documents, team, updates, inspection) ตอบ 409 จนกว่าจะบันทึกสัญญาและตั้งค่างานก่อสร้าง
            operationId: recordProjectContract
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ContractInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Project' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/setup:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: การตั้งค่างานก่อสร้างของโครงการ
            description: ยังไม่ตั้งค่า → configured = false และ options เป็นค่าเริ่มต้น; 409 ถ้ายังไม่บันทึกสัญญา
            operationId: getProjectSetup
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProjectSetup' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
        put:
            tags: [projects]
            summary: บันทึกการตั้งค่างานก่อสร้างและสร้างไทม์ไลน์ (ต้องมีสิทธิ์ project.manage)
            description: |
                หลังบ้านคัดงานจากแม่แบบตามตัวเลือก จัดวันที่ให้พอดีกับระยะสัญญา สร้างไทม์ไลน์และงวดงานใหม่ และบันทึก Audit Log
                ตัวเลือกที่ถูกซ่อน (visibleWhen ไม่ตรง) หลังบ้านปรับเป็นค่าเริ่มต้นเอง
                409 ถ้ายังไม่บันทึกสัญญา หรือเริ่มรายงานความคืบหน้าแล้ว (locked)
            operationId: saveProjectSetup
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ProjectSetupInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProjectSetup' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/setup/preview:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        post:
            tags: [projects]
            summary: ดูตัวอย่างงานที่จะเกิดขึ้นจากตัวเลือก (ไม่บันทึก)
            operationId: previewProjectSetup
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ProjectSetupInput' }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/SetupPreview' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/timeline:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: ไทม์ไลน์ตามแผนงานพร้อมความคืบหน้ารายงาน
            operationId: getProjectTimeline
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProjectTimeline' }
                '404': { $ref: '#/components/responses/NotFound' }
    /projects/{code}/updates:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: บันทึกหน้างาน/ประวัติการอัปเดตความคืบหน้า (ล่าสุดก่อน)
            operationId: listProgressUpdates
            parameters:
                - $ref: '#/components/parameters/Page'
                - $ref: '#/components/parameters/PageSize'
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProgressUpdatePage' }
                '404': { $ref: '#/components/responses/NotFound' }
        post:
            tags: [projects]
            summary: บันทึกความคืบหน้าประจำวัน
            description: |
                หลังบ้านปรับ % ของงานที่ระบุ แล้วคำนวณ % ของขั้นตอนและภาพรวมโครงการใหม่ (ถ่วงน้ำหนักตามจำนวนวันของงาน)
                หมุดหมาย (milestone) เสร็จอัตโนมัติเมื่องานอื่นในขั้นตอนเสร็จครบ
                งานที่เป็นจุดตรวจ (Hold Point) อัปเดตผ่าน endpoint นี้ไม่ได้ ต้องใช้ /inspection
                การลด % ต้องระบุเหตุผลใน note
            operationId: createProgressUpdate
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ProgressUpdateInput' }
            responses:
                '201':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProgressUpdate' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/tasks/{taskCode}/inspection:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
            - { name: taskCode, in: path, required: true, schema: { type: string, examples: ['05.04'] } }
        post:
            tags: [projects]
            summary: บันทึกผลตรวจจุด Hold Point (ผ่าน = งานเสร็จ, ไม่ผ่าน = ต้องแก้ไขและตรวจใหม่)
            operationId: recordInspection
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/InspectionInput' }
            responses:
                '201':
                    description: บันทึกแล้ว (คืนรายการบันทึกหน้างานที่สร้างขึ้น)
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProgressUpdate' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
...
    /projects/{code}/change-orders:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: งานเพิ่ม-ลดของโครงการ (ล่าสุดก่อน)
            operationId: listChangeOrders
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/ChangeOrder' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
        post:
            tags: [projects]
            summary: ขอเพิ่ม-ลดงาน (ต้องมีสิทธิ์ project.manage) — สร้างคำขออนุมัติประเภท change-order ในศูนย์อนุมัติ
            description: |
                อนุมัติ/ไม่อนุมัติผ่าน POST /approvals/{approvalId}/approve|reject เมื่ออนุมัติ หลังบ้าน:
                - บวกยอดสุทธิเข้า changeOrderTotal ของโครงการ และเพิ่มงวดงานของงานเพิ่ม-ลด
                - เลื่อนกำหนดส่งมอบตาม scheduleImpactDays
                - เพิ่ม newTask เข้าไทม์ไลน์ (ท้ายขั้นตอนที่ระบุ ก่อนหมุดหมาย)
                งานที่ลูกค้าขอ (source = customer) อนุมัติไม่ได้จนกว่าลูกค้าจะยืนยัน (409)
            operationId: createChangeOrder
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/ChangeOrderInput' }
            responses:
                '201':
                    description: ส่งขออนุมัติแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ChangeOrder' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/change-orders/{id}/customer-confirm:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [projects]
            summary: บันทึกว่าลูกค้ายืนยันรายการและราคาแล้ว (คำขอที่ยังรออนุมัติ)
            operationId: confirmChangeOrderByCustomer
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ChangeOrder' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /projects/{code}/change-orders/{id}/cancel:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [projects]
            summary: ยกเลิกคำขอที่ยังรออนุมัติ (คำขออนุมัติที่ผูกไว้ถูกปิดด้วย)
            operationId: cancelChangeOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                reason: { type: string }
            responses:
                '200':
                    description: ยกเลิกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ChangeOrder' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /projects/{code}/installments:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: งวดงานและการชำระเงินของลูกค้า
            operationId: listProjectInstallments
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/Installment' }
                '404': { $ref: '#/components/responses/NotFound' }
    /projects/{code}/photos:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            summary: ภาพถ่ายหน้างาน (ล่าสุดก่อน)
            operationId: listProjectPhotos
            parameters:
                - { name: phaseCode, in: query, schema: { type: string } }
                - $ref: '#/components/parameters/Page'
                - $ref: '#/components/parameters/PageSize'
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/SitePhotoPage' }
                '404': { $ref: '#/components/responses/NotFound' }
    /projects/{code}/documents:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [projects]
            operationId: listProjectDocuments
            parameters:
                - { name: category, in: query, schema: { $ref: '#/components/schemas/DocumentCategory' } }
                - { name: q, in: query, schema: { type: string } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/ProjectDocument' }
                '404': { $ref: '#/components/responses/NotFound' }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
    responses:
...
        Forbidden:
            description: บทบาทนี้ไม่มีสิทธิ์
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
...
        ValidationError:
            description: ข้อมูลไม่ถูกต้อง
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
    schemas:
...
        ProjectStatus:
            type: string
            description: pending-contract = เปิดโครงการแล้วแต่ยังไม่ได้บันทึกสัญญา (ยังไม่มีแผนงาน งวดงาน ภาพ เอกสาร)
            enum: [pending-contract, planning, in-progress, near-handover, completed, delayed]
...
        DocumentCategory:
            type: string
            enum: [contract, drawing, permit, inspection, billing, handover]
...
        ProjectGroup:
            type: string
            description: |
                pending-contract = รอทำสัญญา, active = กำลังดำเนินการ (มีสัญญาแล้ว ยังไม่ส่งมอบ),
                in-hand = โครงการในมือ (รอทำสัญญา + กำลังดำเนินการ), completed = ส่งมอบแล้ว,
                warranty = ส่งมอบแล้วและยังอยู่ในระยะประกัน
            enum: [in-hand, pending-contract, active, completed, warranty]
...